Refactor ModelStub to be standalone class
